Use a single unified part for all saves in IronMan.
Use a single part to control all saving logic, rather than individual
parts for individual use cases. This simplifies the mod a bit and makes
it easier to read, and is mildly more efficient than trying to add a new
part for every single imaginable case we might want to deal with down
the road.

Bump IronMan version to 0.3.0.