Stop using dictionaries to store previous save turns
Instead of having custom logic for each event to determine whether or
not to perform a save, we keep a single field with the last time
IronManSavePart was used to trigger a save. In addition, we use a single
MinTurns field to determine how many turns to wait between saves, rather
than a dictionary.